+++ next-docs/components/VisionCard/types/CardBackProps.ts
@@ -1,0 +1,7 @@
+type CardBackProps = {
+  heading: string;
+  subHeading: string;
+  supportingTextItems: string[];
+};
+
+export default CardBackProps;