+++ next-docs/components/breadcrumbs/Breadcrumbs.tsx
@@ -1,3 +1,4 @@
+import { ArrowsRight } from '@heathmont/moon-icons-tw';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 
@@ -9,7 +10,7 @@ const Breadcrumbs = () => {
   }
   return (
     <nav aria-label="Breadcrumb">
-      <ol className="flex flex-wrap pb-12 gap-2 items-center text-moon-12">
+      <ol className="flex flex-wrap lg:pb-12 gap-2 items-center text-moon-12">
         <li className="flex items-center gap-2 text-trunks transition-colors duration-200 hover:text-bulma">
           <Link href="/">
             <a>Moon.io</a>
@@ -21,20 +22,7 @@ const Breadcrumbs = () => {
               key={page + index}
               className="flex items-center gap-2 text-trunks transition-colors duration-200 hover:text-bulma"
             >
-              <svg
-                width="16"
-                height="16"
-                viewBox="0 0 32 32"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M6 16.004L25.996 16m0 0l-5.383-5.386M25.996 16l-5.383 5.385"
-                  stroke="currentColor"
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                ></path>
-              </svg>
+              <ArrowsRight className="text-moon-16 text-trunks" />
               <a href={page}>{page && page[0].toUpperCase() + page.slice(1)}</a>
             </li>
           ))}