+++ next-docs/public/examples/avatarTW/Colours.tsx
@@ -1,0 +1,11 @@
+import { Avatar } from '@heathmont/moon-core-tw';
+
+const Example = () => (
+  <>
+    <Avatar />
+    <Avatar color="text-goten" backgroundColor="bg-piccolo" />
+    <Avatar color="text-goten" backgroundColor="bg-chiChi" />
+  </>
+);
+
+export default Example;