+++ workspaces/core/src/avatar/private/types/SizeProps.ts
@@ -1,0 +1,3 @@
+type SizeProps = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl';
+
+export default SizeProps;