+++ workspaces/core/src/avatar/private/types/StatusOriginProps.ts
@@ -1,0 +1,6 @@
+type StatusOriginProps = {
+  vertical: 'top' | 'bottom';
+  horizontal: 'left' | 'right';
+};
+
+export default StatusOriginProps;